♻️ some refactor
