+++ src/app/(actions)/middlewares.ts
@@ -1,12 +1,24 @@
 import "server-only";
 import { revalidatePath } from "next/cache";
 import { getSession, getAuthedUser } from "~/app/(actions)/auth";
-import type {
-  AuthState,
-  AuthedServerActionResult,
-  OmitFirstItemInArray,
-  AuthError
-} from "~/lib/types";
+import type { OmitFirstItemInArray } from "~/lib/types";
+import type { Session } from "~/lib/server/session.server";
+import type { User } from "~/lib/server/db/schema/user.sql";
+
+export type AuthState = {
+  currentUser: User;
+  session: Session;
+};
+
+export type AuthError = {
+  type: "AUTH_ERROR";
+};
+
+export type AuthedServerActionResult<
+  Action extends (auth: AuthState, ...args: any[]) => Promise<any>
+> = (
+  ...args: OmitFirstItemInArray<Parameters<Action>>
+) => Promise<Awaited<ReturnType<Action>> | AuthError>;
 
 export function withAuth<
   T extends (auth: AuthState, ...args: any[]) => Promise<any>
